AmenityController

+++ src/HostelFinder.Domain/Entities/RoomDetails.cs
@@ -9,7 +9,7 @@ public class RoomDetails
     public int BedRooms { get; set; } 
     public int BathRooms { get; set; }
     public int Kitchen { get; set; }
-    public decimal Size { get; set; }
+    public int Size { get; set; }
     public bool Status { get; set; }
     [MaxLength(255)]
     public string? OtherDetails { get; set; }
